chore: extract common call expression utils
This adds two new utils:

- `findDefaultImportIdentifier` to find imports and a name for a given
	module
- `computeSimpleCallReplacementEdits` to do a basic `foo($ARGS)` ->
	`bar($ARGS)` edit (i.e. just changing the name)
